+++ NEWS.md
@@ -1,6 +1,7 @@
 # NEWS
 
 ## Version 2.1.3
+* Added package vignette
 * The custom argument `force.pairwise` used in some of the `proxy` distances is not necessary anymore
 * Added Silhouette Cluster Validity Index
 * No longer enforcing `NULL` dimensions for each series (in case multivariate series are to be used)