+++ NEWS.md
@@ -1,6 +1,6 @@
 # NEWS
 
-## Version 2.1.3
+## Version 2.2.0
 * Added package vignette
 * Implemented several cluster validity indices in the new `cvi` function
 * The custom argument `force.pairwise` used in some of the `proxy` distances is not necessary anymore